+++ src/DbReader/DbReader.csproj
@@ -34,7 +34,7 @@
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="LightInject.Source" Version="6.2.0">
+    <PackageReference Include="LightInject.Source" Version="6.3.0">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
     <!-- <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All"/> -->
