+++ src/DbReader/DynamicArguments/IDynamicType.cs
@@ -1,0 +1,17 @@
+namespace DbReader.DynamicArguments
+{
+    /// <summary>
+    /// Implemented by the types generated by the `ArgumentsBuilder` class to be able to get an value by name.
+    /// </summary>
+    public interface IDynamicType
+    {
+        /// <summary>
+        /// Gets the value from an <see cref="IDynamicType"/> by name.
+        /// </summary>
+        /// <param name="memberName">The name of the member for which to get the value.</param>
+        /// <typeparam name="T">The type of value to get.</typeparam>
+        /// <returns>The value of the member.</returns>
+        T Get<T>(string memberName);
+    }
+
+}
